+++ backend/NumberInformation.API/NumberInformation.Tests/ProcessBackgroundTest.cs
@@ -18,14 +18,14 @@ namespace NumberInformation.Tests
             _processBackground = new MemoryProcessBackground<NIResponse>(); 
         }
 
-        [TestCase(10,new int[] { 1, 2, 5, 10 }, new int[] { 1, 2, 5 }) ]
-        [TestCase(100, new int[] { 1, 2, 4, 5, 10, 20, 25, 50, 100 }, new int[] { 1, 2, 5 })]
-        [TestCase(13, new int[] { 1, 13 }, new int[] { 1, 13 })]
-        [TestCase(15, new int[] { 1, 3, 5, 15 }, new int[] { 1, 3, 5 })]
-        [TestCase(24, new int[] { 1, 2, 3, 4, 6,  8, 12, 24  }, new int[] { 1, 2, 3})]
-        [TestCase(999983, new int[] { 1, 999983 }, new int[] { 1, 999983 })]
+        [TestCase(10,new long[] { 1, 2, 5, 10 }, new long[] { 1, 2, 5 }) ]
+        [TestCase(100, new long[] { 1, 2, 4, 5, 10, 20, 25, 50, 100 }, new long[] { 1, 2, 5 })]
+        [TestCase(13, new long[] { 1, 13 }, new long[] { 1, 13 })]
+        [TestCase(15, new long[] { 1, 3, 5, 15 }, new long[] { 1, 3, 5 })]
+        [TestCase(24, new long[] { 1, 2, 3, 4, 6,  8, 12, 24  }, new long[] { 1, 2, 3})]
+        [TestCase(999983, new long[] { 1, 999983 }, new long[] { 1, 999983 })]
         
-        public void DividingNumber_Cases( int inputNumber, int[] dividingNumbers, int[] primeNumbers)
+        public void DividingNumber_Cases(long inputNumber, long[] dividingNumbers, long[] primeNumbers)
         {
             var taskId = _processBackground.ExecuteInBackground(() => inputNumber.GetDividingInformations());
             var result = _processBackground.GetResultBackgroundTask(taskId);
@@ -35,8 +35,8 @@ namespace NumberInformation.Tests
             }
 
             result.Input.ShouldBe(inputNumber);
-            result.DividingNumbers.ShouldBe(new List<int>(dividingNumbers));
-            result.DividingPrime.ShouldBe(new List<int>(primeNumbers));
+            result.DividingNumbers.ShouldBe(new List<long>(dividingNumbers));
+            result.DividingPrime.ShouldBe(new List<long>(primeNumbers));
         }
 
     }
